Clean up main file
